faet: view product ui
